bezier curve finder tool
begin by copying the track.JPG to the parent directory.  then run
program
click end points of the curve along the track.  the last point will be
the first point.
after all end points have been added, then add control points that best
fit the track and how you would like your kart to move around the
underlying track.  (right click to add to the collection of control
points)
once finished a print out of all of the collected points will appear in
the processing console. this is the input for the race track.